"Add email sending functionality in application
Added functionality to send emails using nodemailer in the application. Created an emailSender utility that uses SMTP connection for sending emails. It defines parameters for the email and uses SMTP credentials from the environment variables.

Also updated existing EmailForm component to use this new utility and send a POST request to the server when form is submitted. It includes error handling and form reset upon successful email send.

Dependency packages for nodemailer and dotenv were installed. Types for nodemailer were added as well."