Improve lang switch functionality for Hebrew language
This update revises the language switch functionality to better support the Hebrew language. The LangSwitch component now accepts a 'lang' prop to determine the currently active language. The SVG path logic has been modified in the motion.path of LangSwitch component to consider directionality ('rtl' for Hebrew). Hebrew language now accurately triggers the 'isOpen' state in LangSwitch. The Header component was also updated to pass the 'lang' prop to LangSwitch. This enhancement is crucial to ensuring smooth user experience for Hebrew language users, taking into account that Hebrew is a right-to-left language.