Update language handling for major components
The props of components were enhanced by adding a language prop. With this addition, all major components are now able to handle language changes more effectively, especially for languages that have different directionality, like Hebrew. This modification was strictly necessary due to the absolute need for supporting multiple languages in a way that conforms to each language's unique characteristics.