Refactor image imports to streamline code
Significant changes have been made to how images are imported in the WorkGrid component. Direct imports of images have been removed and rewired through the ‘data.ts’ file in a convenient object form. This reduction in direct import usage not only makes the WorkGrid cleaner and more readable, but it also optimizes code by marshalling resources in a centralized manner. This change does not affect function, but enhances code quality and maintainability. Meanwhile, a slight adjustment has been made to the key property in the SwiperSlide component to include a hyphen for better readability.