Add new SVG icon and refactor folder structure for easier localization
A new SVG icon has been added to the application as part of ongoing design improvements. File structure for 'about', 'projects', and 'services' pages have been refactored in preparation for a future internationalization feature. This includes introducing a '[lang]' directory to house different language versions of the pages, contributing to a more organized codebase and easier file navigability.