Improve ContactErrorMessage UI and add success notification
This commit updates the position and layout of the ContactErrorMessage component to centrally align it for a better UI. An additional 'successMessage' prop is also introduced for the component to handle successful email submissions. Functional state logic has been added to Footer.tsx to account for error notifications and success messages, which are triggered upon form submissions. The EmailForm.tsx has been refactored to handle success notification as well. These changes aim to enhance user interaction and feedback upon form submissions.