Add redirect rule in Next.js configuration
This commit introduces a redirect rule in the Next.js configuration file (next.config.js). Now, users navigating to the root URL ("/") will be automatically redirected to the "/en" path. The change ensures a unified and consistent user experience across the site by directing all users to the English version of the site by default.