Handle root path in locale redirect middleware
Previously, the logic of our locale redirect middleware did not consider the scenario when a user hits the root path of the app ('/'). That could lead to an undesired user experience as no language redirection was made.

In this commit, the middleware was updated to consider the root path a 'missing locale' and thus, the language redirection is now also triggered when the user enters the application through the root path. This ensures that users are always served content in their preferred locale.