Add Vercel analytics to the application
Vercel analytics was added to the application layout in order to better understand our user behavior and track website usage. Dependency for '@vercel/analytics' was added to the package.json and package-lock.json files. This was then imported into our application within the layout.tsx file. The `<Analytics />` component is now appended as the last child in HTML `<body>` element, allowing to track every page rendered.