Refactor dictionaries.ts and page.tsx for optimal operation
These changes refactor the function getDictionary in dictionaries.ts and sections importation in page.tsx. The logic in getDictionary that throws errors for unsupported locale was commented out. This allows for flexibility during testing of different locales. Note, this should be revised or removed entirely in subsequent updates. The order of importing sections in page.tsx has been revised. It's expected to improve the efficiency of the code and make the user interface more effective.